PushUpdates

+++ FPDF/print/mc_table.php
@@ -247,7 +247,7 @@ class PDF_MC_Table extends FPDF
             $item_body = queryItem("call get_body_section('$item->id','$section_id')")[0];
 
         $item_body->qty = $item_body->jan + $item_body->feb + $item_body->mar + $item_body->apr + $item_body->may + $item_body->jun + $item_body->jul + $item_body->aug + $item_body->sep + $item_body->oct + $item_body->nov + $item_body->dece;
-        $item_body->estimated_budget = ((int)$item_body->qty * str_replace(',', '', (int)$item_body->unit_cost));
+        $item_body->estimated_budget = ((int)$item_body->qty * str_replace(',', '', (float)$item_body->unit_cost));
 
         $sum = 0;
         if($item->expense_id == "1")
